+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Move `default_due` config to core
+- Move `default_priority` config to core
 
 ## [v0.2.2] - 2025-07-28